brain/requirements.txt
mcap>=1.1
mcap-ros2-support>=0.5
pandas>=2.0
numpy>=1.24
scipy>=1.10
